\name{ASCII_MSDataCreation}

\alias{ASCII_MSDataCreation}

\docType{data}
\title{
Demonstration folder for MS.DataCreation with option DataType=ASCII
}

\description{
This demonstration folder includes 2 transformed GC-MS analyses of Petrel obtained from trans.ASCII. The two analyses represent a total of 67 chromatogram's peaks. 
The folder can be used with the function MS.DataCreation that collects and assembles the information from chromatograms and mass spectra of the two samples in a initial data matrix with peaks in row and mass spectrum in columns.  
}

\usage{ASCII_MSDataCreation}

\format{ A folder with two different transformed ascii files, each corresponding to one GC-MS analysis. 
 }

\examples{
data(ASCII_MSDataCreation)
}

\keyword{datasets}
