\name{ASCII_TransASCII}

\alias{ASCII_TransASCII}

\docType{data}
\title{
Demonstration folder for trans.ASCII
}

\description{
This demonstration folder includes 2 raw GC-MS analyses of Petrel in ASCII format. 
The data in ASCII format have to be transformed with the function trans.ASCII for further analyses with MS.DataCreation. 
The folder can be used with the function trans.ASCII to transform the raw ascii GC-MS data in the format suitable for MS.DataCreation.  
}

\usage{ASCII_TransASCII}

\format{ A folder with two different raw ascii files corresponding to the two different GC-MS analyses. 
 }

\examples{
pathASCII<-system.file("doc/ASCII_TransASCII",package="MSeasy")
}

\keyword{datasets}
